fix: correct uint32 handling in DecodeUtils and EncodeUtils
- Updated readUInt32 method to mask the result with 0xFFFF_FFFFL for proper uint32 decoding.
- Modified writeUInt32 method to accept long values directly, ensuring correct encoding of uint32 values.
- Added boundary tests for readUInt32 and writeUInt32 methods to validate min/max values and error handling for out-of-range inputs.

+++ src/main/java/org/indunet/fastproto/util/EncodeUtils.java
@@ -306,7 +306,7 @@ public class EncodeUtils {
      * @param value The uint32 value to be encoded.
      */
     public static void writeUInt32(byte[] bytes, int offset, long value) {
-        writeUInt32(bytes, offset, ByteOrder.LITTLE, (int) value);
+        writeUInt32(bytes, offset, ByteOrder.LITTLE, value);
     }
 
     /**
